[동호]데이터 수정
