+++ src/pages/commerce/FilterCategory.jsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import FilterComp from "../components/detail/FilterComp";
+import FilterComp from "../../components/detail/FilterComp";
 
 const FilterCategory = () => {
   const { categoryName } = useParams();
