+++ src/pages/commerce/Star.jsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import Masonry from "react-masonry-css";
-import { StarData } from "../StarData";
+import { StarData } from "../../StarData";
 import Pagination from "react-js-pagination";
 
 const Container = styled.div`
