+++ src/pages/commerce/eventpages/Influencer.jsx
@@ -1,23 +1,23 @@
 import React from "react";
-import ScrollButton from "../../components/event/ScrollButton";
+import ScrollButton from "../../../components/event/ScrollButton";
 import styled from "styled-components";
-import InfluencerBanner from "../../components/event/InfluencerBanner";
-import isabe01 from "../../imgs/event/isabe01.jpg";
-import isabe02 from "../../imgs/event/isabe02.jpg";
-import isabe03 from "../../imgs/event/isabe03.jpg";
-import isabe04 from "../../imgs/event/isabe04.jpg";
-import jungwon01 from "../../imgs/event/jungwon01.jpg";
-import jungwon02 from "../../imgs/event/jungwon02.jpg";
-import jungwon03 from "../../imgs/event/jungwon03.jpg";
-import leoj01 from "../../imgs/event/leoj01.jpg";
-import leoj02 from "../../imgs/event/leoj02.jpg";
-import leoj03 from "../../imgs/event/leoj03.jpeg";
-import leoj04 from "../../imgs/event/leoj04.webp";
-import isabeprofile from "../../imgs/event/isabeprofile.jpg";
-import jungwonprofile from "../../imgs/event/jungwonprofile.webp";
-import leojprofile from "../../imgs/event/leojprofile.webp";
-import blurimg from "../../imgs/event/blurimage.png";
-import blurimg02 from "../../imgs/event/blurimage02.png";
+import InfluencerBanner from "../../../components/event/InfluencerBanner";
+import isabe01 from "../../../imgs/event/isabe01.jpg";
+import isabe02 from "../../../imgs/event/isabe02.jpg";
+import isabe03 from "../../../imgs/event/isabe03.jpg";
+import isabe04 from "../../../imgs/event/isabe04.jpg";
+import jungwon01 from "../../../imgs/event/jungwon01.jpg";
+import jungwon02 from "../../../imgs/event/jungwon02.jpg";
+import jungwon03 from "../../../imgs/event/jungwon03.jpg";
+import leoj01 from "../../../imgs/event/leoj01.jpg";
+import leoj02 from "../../../imgs/event/leoj02.jpg";
+import leoj03 from "../../../imgs/event/leoj03.jpeg";
+import leoj04 from "../../../imgs/event/leoj04.webp";
+import isabeprofile from "../../../imgs/event/isabeprofile.jpg";
+import jungwonprofile from "../../../imgs/event/jungwonprofile.webp";
+import leojprofile from "../../../imgs/event/leojprofile.webp";
+import blurimg from "../../../imgs/event/blurimage.png";
+import blurimg02 from "../../../imgs/event/blurimage02.png";
 import { useNavigate } from "react-router-dom";
 
 const PageWrapper = styled.div`
