+++ src/pages/ott/Original.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Original = () => {
+  return <div>Original</div>;
+};
+
+export default Original;
